requirements.txt
fastapi==0.115.12
multipart==1.2.1
python-dotenv==1.0.1